Display Request Changes
- Change 'infoQualify' data type from 'String' to 'InfoQualify'
- Create separate classes for 'OutputContent' and 'OutputText' to be
consistent with other models
- Add 'private' access modifier to 'displayOutput'
