+++ src/main/java/au/com/dmg/fusion/request/displayrequest/OutputText.java
@@ -23,41 +23,19 @@
 
 package au.com.dmg.fusion.request.displayrequest;
 
-import au.com.dmg.fusion.util.InstantAdapter;
 import com.squareup.moshi.Json;
-import com.squareup.moshi.JsonAdapter;
-import com.squareup.moshi.Moshi;
 
-import au.com.dmg.fusion.util.BigDecimalAdapter;
-import au.com.dmg.fusion.request.Request;
+public class OutputText {
 
-public class DisplayRequest implements Request {
+	@Json(name = "Text")
+	private final String text;
 
-	@Json(name = "DisplayOutput")
-	private DisplayOutput displayOutput;
-
-	public DisplayRequest() {
-	}
-	
-	public DisplayRequest(DisplayOutput displayOutput) {
-		this.displayOutput = displayOutput;
+	public OutputText(String text) {
+		this.text = text;
 	}
 
-    public DisplayOutput getDisplayOutput() {
-		return displayOutput;
+	public String getText() {
+		return text;
 	}
 
-	public void setDisplayOutput(DisplayOutput displayOutput) {
-		this.displayOutput = displayOutput;
-	}
-
-	@Override
-    public String toJson() {
-        Moshi moshi = new Moshi.Builder()
-                .add(new BigDecimalAdapter())
-				.add(new InstantAdapter())
-                .build();
-        JsonAdapter<DisplayRequest> jsonAdapter = moshi.adapter(DisplayRequest.class);
-        return jsonAdapter.toJson(this);
-    }
 }
